% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_htn.R
\name{correct_htn}
\alias{correct_htn}
\title{Prevalence of hypertension (HT) corrected from sampling variability}
\usage{
correct_htn(
  form = htn ~ 1,
  subpop = NULL,
  n_samp = 1000,
  data_long = dt_nhanes,
  surv_des = NULL,
  correct = TRUE,
  tresh = dplyr::tibble(BP_typ = c("sys", "dia"), BP_tresh = c(140, 90))
)
}
\arguments{
\item{form}{A formula giving covariates by which HT prevalence is
calculated. If \code{htn} is placed in the left hand side of
\code{formula}, then variation of prevalence according to the
covariates given in the right hand side of are estimated with a
\code{glm} (\code{survey::svyglm} if \code{surv_des} is not NULL) model with
quasibinomial distribution. Else, prevalence is tabulated
according to the covariates given in the right hand side (simple
tabulation or \code{survey::svyby} if \code{surv_des} is not NUL).}

\item{subpop}{A boolean covariate that defines a subpopulation
over which to filter the calculation of prevalence. \code{subpop}
must be given in a formula form, e.g. ~subpop. See Examples.}

\item{n_samp}{The number of posterior sample of the correction
factor to be used in the estimation. Default (NULL) resumes to
the maximum number of available posterior samples.}

\item{data_long}{Data frame giving the BP measurements of the
population under study. Data must be given in a long format,
e.g. one raw per BP measure (see \code{dt_nhanes} format).  The data
frame must have the following columns: \itemize{
\item{"id"}{Patient identifyer} \item{"age"}{Age of the patient, in years}
\item{"sex"}{Sex of the patient} \item{"tt_htn"}{Boolean with
value \code{TRUE} if the patient is under anti-hypertensive
treatment} \item{"visit"}{Identifyer of the visit}
\item{"BP_typ"}{type of BP prelevement, "dia" for diastolic,
"sys" for systolic} \item{"bp"}{Value of blood pressure
measurement} }}

\item{surv_des}{If applicable, the survey design of the study,
specifyed with \code{svydesign} from \code{survey} package.}

\item{correct}{Boolean set to \code{TRUE} (the default) to correct
prevalence.}

\item{tresh}{A data frame giving the BP thresholds that defines
hypertension. Default to 140 for systolic BP, 90 for diastolic.}
}
\value{
A data frame with the estimates of HT prevalance tabulated according to the covariates given in \code{form}
}
\description{
This function allows to calculate the prevalence of HT in a population
after correcting from variations due to blood pressure variability that occurs in an indivuidual
between visits and within visit.
For each individual, a correction factors adapted to scheme of BP prelevement (i.e. number of visits and number of measure within a visit, systolic or diastolic measure), age and sex, is calculated. These correction factor are derived from BP variances estimted from NHANES III datas. The corrections factors are then applied to indivuduals BP measures in the population under study to infer corrected HT prevalence.
}
\examples{
data(dt_nhanes)

## Uncorrected HT prevalence by sex
correct_htn(form= ~ sex,
data_long = dt_nhanes,
correct = FALSE)

## Corrected HT prevalence by sex
correct_htn(form=htn ~ sex,
data_long = dt_nhanes,
n_samp = 10,
correct = TRUE)

## Same using quasi-binomial glm
correct_htn(form= ~ sex,
data_long = dt_nhanes,
n_samp = 10,
correct = TRUE)

}
\author{
Edouard Chatignoux
}
